requirements-minimal.txt
# Core
flask==2.3.3
flask-login==0.6.2
werkzeug==2.3.7
python-dotenv==1.0.0

# Google Cloud
google-cloud-firestore==2.15.0
google-auth==2.23.4
google-api-core==2.15.0
google-generativeai==0.3.2

# Database
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0

# Utilities
python-dateutil==2.8.2
typing-extensions==4.8.0